Pulls out Scanner from RecordStore
